+++ Model/DeletedData.php
@@ -4,7 +4,7 @@ namespace Leantime\Plugins\APIData\Model;
 
 use Carbon\CarbonInterface;
 
-class DeletedData
+readonly class DeletedData
 {
     public function __construct(
         public ?int $id,